requirements.txt
Django==4.2.10
djangorestframework==3.14.0
pandas==2.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.6
gunicorn==21.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
